refactor gsutil and support withGCPEnv
